requirements.txt
# GeoClaw — core + intelligence stack (Tasks 1–7)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.36
requests>=2.31.0
apscheduler>=3.10.0
